Show add manga dropdown after user clicks on a button
There is a bug in streamlit where if you click on a button (on the
sidebar) to show a dialog, the page has to rerun before clicking on the
button again, or else the dialog doesn't open. Forcing the user to click
on another button before clicking on the button to open the dialog
"solves" it.